Move type definition to centralised types
